message handler spec moved
